chore(schema-files-loader): extracted types previously used in internals
